+++ Gears/ScriptSettings.hpp
@@ -290,6 +290,11 @@ public:
         } Native;
     } Controller;
 
+    // [MISC]
+    struct {
+        bool UDPTelemetry = true;
+    } Misc;
+
     // [UPDATE]
     struct {
         bool EnableUpdate = true;